LanguageManager
I made a new swift file that deals with checks the language of the app and device. It will check if the languages are different and if they are, the app will prompt the onboarding language selection page again. Then, you can change your language to the one that matches your device.